Create session token for user upon signup
