+++ package.json
@@ -13,7 +13,7 @@
     "@prisma/client": "^5.7.0",
     "bcrypt": "^5.1.1",
     "fastify": "^3.21.5",
-    "nanoid": "^5.0.4",
+    "nanoid": "^3.3.4",
     "prisma-multischema": "^1.1.4"
   },
   "devDependencies": {
