+++ src/models/signupRequestModel.ts
@@ -1,4 +1,5 @@
 export interface SignupRequestModel {
     username: string;
     plainPassword: string;
+    browserInfo: string
 }
